Integration Tests were renamed.
