+++ src/test/scala-2.11/agents/integration/ProducerAndConsumerLazyTest.scala
@@ -9,7 +9,7 @@ import testutils._
 import scala.util.control.Breaks._
 
 
-class AProducerAndConsumerLazyTest extends FlatSpec with Matchers with BeforeAndAfterAll with TestUtils {
+class ProducerAndConsumerLazyTest extends FlatSpec with Matchers with BeforeAndAfterAll with TestUtils {
 
   f.setProperty(TSF_Dictionary.Stream.NAME,"test_stream").
     setProperty(TSF_Dictionary.Stream.PARTITIONS,3).
