+++ src/test/scala-2.11/agents/integration/SubscriberV2BasicFunctions.scala
@@ -1,0 +1,8 @@
+package agents.integration
+
+/**
+  * Created by ivan on 24.08.16.
+  */
+class SubscriberV2BasicFunctions {
+
+}
